Update code
Due to deprecation (since Python 3.1) json.load has since Python 3.9 no longer the ability for declaring an encoding as parameter. A quick fix would be to simply remove this parameter (as done here). This could cause problems for example when using Greek letters. Furthermore it would be nice if you provide release numbers/tags in the releases section of GitHub.